pyproject.toml
[project]
name = "deeplearning1-classification"
version = "0.1.0"
description = "Config-driven multi-task text classification pipeline"
readme = "README.md"
authors = [{ name = "Project Team" }]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24",
    "pandas>=1.5",
    "scikit-learn>=1.2",
    "torch>=2.1",
    "transformers>=4.35",
    "optuna>=3.1",
    "PyYAML>=6.0",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
develop = [
    "black>=23.0",
    "ruff>=0.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"
